+++ packages/expo-router/src/modal/Modal.tsx
@@ -2,11 +2,12 @@
 
 import { type NavigationProp, type ParamListBase } from '@react-navigation/native';
 import { nanoid } from 'nanoid/non-secure';
-import { useEffect, useState } from 'react';
-import { StyleSheet, ViewProps } from 'react-native';
+import { use, useEffect, useState } from 'react';
+import { View, StyleSheet, type ViewProps } from 'react-native';
 import { type ScreenProps } from 'react-native-screens';
 
 import { useModalContext, type ModalConfig } from './ModalContext';
+import { ModalPortalContent, PortalContentHeightContext } from './Portal';
 import { useNavigation } from '../useNavigation';
 import { areDetentsValid } from './utils';
 
@@ -141,13 +142,12 @@ export function Modal(props: ModalProps) {
     if (visible) {
       const newId = nanoid();
       openModal({
+        component: children,
         animationType,
         presentationStyle,
         transparent,
         viewProps,
-        component: children,
         uniqueId: newId,
-        parentNavigationProp: navigation,
         detents: detents ?? (presentationStyle === 'formSheet' ? 'fitToContents' : undefined),
       });
       setCurrentModalId(newId);
@@ -197,5 +197,29 @@ export function Modal(props: ModalProps) {
     }
     return () => {};
   }, [currentModalId, addEventListener, onClose, onShow]);
-  return null;
+  if (!currentModalId || !visible) {
+    return null;
+  }
+  return (
+    <ModalPortalContent hostId={currentModalId}>
+      <ModalContent {...viewProps}>{children}</ModalContent>
+    </ModalPortalContent>
+  );
+}
+
+function ModalContent(props: ViewProps) {
+  const { children, ...viewProps } = props;
+  const { setHeight } = use(PortalContentHeightContext);
+  return (
+    <View
+      {...viewProps}
+      onLayout={(e) => {
+        const { height } = e.nativeEvent.layout;
+        if (height) {
+          setHeight(height);
+        }
+      }}>
+      {children}
+    </View>
+  );
 }