+++ packages/expo-router/CHANGELOG.md
@@ -39,6 +39,7 @@
 - Fix modal with detents fixToContents on Android ([#38440](https://github.com/expo/expo/pull/38440) by [@Ubax](https://github.com/Ubax))
 - Fix detents in web modal ([#38446](https://github.com/expo/expo/pull/38446) by [@Ubax](https://github.com/Ubax))
 - Fix Link Preview between two separate stacks ([#38299](https://github.com/expo/expo/pull/38299) by [@Ubax](https://github.com/Ubax))
+- use portal in Modal ([#37980](https://github.com/expo/expo/pull/37980) by [@Ubax](https://github.com/Ubax))
 
 ### 💡 Others